added xnu-792.25.20

+++ bsd/dev/i386/systemcalls.c
@@ -1,29 +1,23 @@
 /*
  * Copyright (c) 2000-2005 Apple Computer, Inc. All rights reserved.
  *
- * @APPLE_OSREFERENCE_LICENSE_HEADER_START@
+ * @APPLE_LICENSE_HEADER_START@
  * 
- * This file contains Original Code and/or Modifications of Original Code
- * as defined in and that are subject to the Apple Public Source License
- * Version 2.0 (the 'License'). You may not use this file except in
- * compliance with the License. The rights granted to you under the License
- * may not be used to create, or enable the creation or redistribution of,
- * unlawful or unlicensed copies of an Apple operating system, or to
- * circumvent, violate, or enable the circumvention or violation of, any
- * terms of an Apple operating system software license agreement.
+ * The contents of this file constitute Original Code as defined in and
+ * are subject to the Apple Public Source License Version 1.1 (the
+ * "License").  You may not use this file except in compliance with the
+ * License.  Please obtain a copy of the License at
+ * http://www.apple.com/publicsource and read it before using this file.
  * 
- * Please obtain a copy of the License at
- * http://www.opensource.apple.com/apsl/ and read it before using this file.
- * 
- * The Original Code and all software distributed under the License are
- * distributed on an 'AS IS' basis, WITHOUT WARRANTY OF ANY KIND, EITHER
+ * This Original Code and all software distributed under the License are
+ * distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY KIND, EITHER
  * EXPRESS OR IMPLIED, AND APPLE HEREBY DISCLAIMS ALL SUCH WARRANTIES,
  * INCLUDING WITHOUT LIMITATION, ANY WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE, QUIET ENJOYMENT OR NON-INFRINGEMENT.
- * Please see the License for the specific language governing rights and
- * limitations under the License.
+ * FITNESS FOR A PARTICULAR PURPOSE OR NON-INFRINGEMENT.  Please see the
+ * License for the specific language governing rights and limitations
+ * under the License.
  * 
- * @APPLE_OSREFERENCE_LICENSE_HEADER_END@
+ * @APPLE_LICENSE_HEADER_END@
  */
 #include <kern/task.h>
 #include <kern/thread.h>
@@ -75,7 +69,7 @@ unix_syscall(x86_saved_state_t *state)
 {
 	thread_t	thread;
 	void		*vt; 
-	unsigned short	code;
+	unsigned int	code;
 	struct sysent	*callp;
 	int		nargs;
 	int		error;
@@ -110,15 +104,14 @@ unix_syscall(x86_saved_state_t *state)
 		/* NOTREACHED */
 	}
 
-	//printf("[scall : eax %x]",  regs->eax);
 	code = regs->eax;
 	params = (vm_offset_t) ((caddr_t)regs->uesp + sizeof (int));
-	callp = (code >= nsysent) ? &sysent[63] : &sysent[code];
+	callp = (code >= (unsigned int)nsysent) ? &sysent[63] : &sysent[code];
 
 	if (callp == sysent) {
 		code = fuword(params);
 		params += sizeof (int);
-		callp = (code >= nsysent) ? &sysent[63] : &sysent[code];
+		callp = (code >= (unsigned int)nsysent) ? &sysent[63] : &sysent[code];
 	}
 	vt = (void *)uthread->uu_arg;
 
@@ -259,7 +252,7 @@ void
 unix_syscall64(x86_saved_state_t *state)
 {
 	thread_t	thread;
-	unsigned short	code;
+	unsigned int	code;
 	struct sysent	*callp;
 	void		*uargp;
 	int		args_in_regs;
@@ -296,7 +289,7 @@ unix_syscall64(x86_saved_state_t *state)
 	args_in_regs = 6;
 
 	code = regs->rax & SYSCALL_NUMBER_MASK;
-	callp = (code >= nsysent) ? &sysent[63] : &sysent[code];
+	callp = (code >= (unsigned int)nsysent) ? &sysent[63] : &sysent[code];
 	uargp = (void *)(&regs->rdi);
 
 	if (callp == sysent) {
@@ -304,8 +297,8 @@ unix_syscall64(x86_saved_state_t *state)
 		 * indirect system call... system call number
 		 * passed as 'arg0'
 		 */
-	        code = regs->rdi;
-		callp = (code >= nsysent) ? &sysent[63] : &sysent[code];
+		code = regs->rdi;
+		callp = (code >= (unsigned int)nsysent) ? &sysent[63] : &sysent[code];
 		uargp = (void *)(&regs->rsi);
 		args_in_regs = 5;
 	}
@@ -468,7 +461,7 @@ unix_syscall_return(int error)
 	thread_t		thread;
 	struct uthread		*uthread;
 	struct proc *p;
-	unsigned short code;
+	unsigned int code;
 	vm_offset_t params;
 	struct sysent *callp;
 	unsigned int cancel_enable;
@@ -485,7 +478,7 @@ unix_syscall_return(int error)
 
 	        /* reconstruct code for tracing before blasting rax */
 		code = regs->rax & SYSCALL_NUMBER_MASK;
-		callp = (code >= nsysent) ? &sysent[63] : &sysent[code];
+		callp = (code >= (unsigned int)nsysent) ? &sysent[63] : &sysent[code];
 
 		if (callp == sysent)
 		        /*
@@ -540,7 +533,7 @@ unix_syscall_return(int error)
 
 		/* reconstruct code for tracing before blasting eax */
 		code = regs->eax;
-		callp = (code >= nsysent) ? &sysent[63] : &sysent[code];
+		callp = (code >= (unsigned int)nsysent) ? &sysent[63] : &sysent[code];
 
 		if (callp == sysent) {
 		        params = (vm_offset_t) ((caddr_t)regs->uesp + sizeof (int));
